+++ pymc3/distributions/continuous.py
@@ -53,12 +53,12 @@ from pymc3.aesaraf import floatX
 from pymc3.distributions import logp_transform, transforms
 from pymc3.distributions.dist_math import (
     SplineWrapper,
+    betainc,
     betaln,
     bound,
     clipped_beta_rvs,
     gammaln,
     i0e,
-    incomplete_beta,
     log_normal,
     logpow,
     normal_lccdf,
@@ -1262,22 +1262,18 @@ class Beta(UnitContinuous):
         Parameters
         ----------
         value: numeric
-            Value(s) for which log CDF is calculated.
+            Value(s) for which log-probability is calculated. If the log probabilities for multiple
+            values are desired the values must be provided in a numpy array or Aesara tensor
 
         Returns
         -------
         TensorVariable
         """
-        # incomplete_beta function can only handle scalar values (see #4342)
-        if np.ndim(value):
-            raise TypeError(
-                f"Beta.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
-            )
 
         return bound(
             at.switch(
                 at.lt(value, 1),
-                at.log(incomplete_beta(alpha, beta, value)),
+                at.log(betainc(alpha, beta, value)),
                 0,
             ),
             0 <= value,
@@ -1937,18 +1933,13 @@ class StudentT(Continuous):
         Parameters
         ----------
         value: numeric
-            Value(s) for which log CDF is calculated.
+            Value(s) for which log-probability is calculated. If the log probabilities for multiple
+            values are desired the values must be provided in a numpy array or Aesara tensor
 
         Returns
         -------
         TensorVariable
         """
-        # incomplete_beta function can only handle scalar values (see #4342)
-        if np.ndim(value):
-            raise TypeError(
-                f"StudentT.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
-            )
-
         lam, sigma = get_tau_sigma(sigma=sigma)
 
         t = (value - mu) / sigma
@@ -1956,7 +1947,7 @@ class StudentT(Continuous):
         x = (t + sqrt_t2_nu) / (2.0 * sqrt_t2_nu)
 
         return bound(
-            at.log(incomplete_beta(nu / 2.0, nu / 2.0, x)),
+            at.log(betainc(nu / 2.0, nu / 2.0, x)),
             0 < nu,
             0 < sigma,
             0 < lam,