+++ pymc3/distributions/discrete.py
@@ -29,11 +29,11 @@ from scipy import stats
 
 from pymc3.aesaraf import floatX, intX, take_along_axis
 from pymc3.distributions.dist_math import (
+    betainc,
     betaln,
     binomln,
     bound,
     factln,
-    incomplete_beta,
     log_diff_normal_cdf,
     logpow,
     normal_lccdf,
@@ -143,24 +143,19 @@ class Binomial(Discrete):
         Parameters
         ----------
         value: numeric
-            Value for which log CDF is calculated.
+            Value(s) for which log-probability is calculated. If the log probabilities for multiple
+            values are desired the values must be provided in a numpy array or Aesara tensor
 
         Returns
         -------
         TensorVariable
         """
-        # incomplete_beta function can only handle scalar values (see #4342)
-        if np.ndim(value):
-            raise TypeError(
-                f"Binomial.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
-            )
-
         value = at.floor(value)
 
         return bound(
             at.switch(
                 at.lt(value, n),
-                at.log(incomplete_beta(n - value, value + 1, 1 - p)),
+                at.log(betainc(n - value, value + 1, 1 - p)),
                 0,
             ),
             0 <= value,
@@ -730,20 +725,15 @@ class NegativeBinomial(Discrete):
         Parameters
         ----------
         value: numeric
-            Value for which log CDF is calculated.
+            Value(s) for which log-probability is calculated. If the log probabilities for multiple
+            values are desired the values must be provided in a numpy array or Aesara tensor
 
         Returns
         -------
         TensorVariable
         """
-        # incomplete_beta function can only handle scalar values (see #4342)
-        if np.ndim(value):
-            raise TypeError(
-                f"NegativeBinomial.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
-            )
-
         return bound(
-            at.log(incomplete_beta(n, at.floor(value) + 1, p)),
+            at.log(betainc(n, at.floor(value) + 1, p)),
             0 <= value,
             0 < n,
             0 <= p,
@@ -1459,18 +1449,13 @@ class ZeroInflatedBinomial(Discrete):
         Parameters
         ----------
         value: numeric
-            Value for which log CDF is calculated.
+            Value(s) for which log-probability is calculated. If the log probabilities for multiple
+            values are desired the values must be provided in a numpy array or Aesara tensor
 
         Returns
         -------
         TensorVariable
         """
-
-        # logcdf can only handle scalar values due to limitation in Binomial.logcdf
-        if np.ndim(value):
-            raise TypeError(
-                f"ZeroInflatedBinomial.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
-            )
 
         return bound(
             logaddexp(at.log1p(-psi), at.log(psi) + _logcdf(binomial, value, {}, n, p)),
@@ -1614,18 +1599,13 @@ class ZeroInflatedNegativeBinomial(Discrete):
         Parameters
         ----------
         value: numeric
-            Value for which log CDF is calculated.
+            Value(s) for which log-probability is calculated. If the log probabilities for multiple
+            values are desired the values must be provided in a numpy array or Aesara tensor
 
         Returns
         -------
         TensorVariable
         """
-        # logcdf can only handle scalar values due to limitation in NegativeBinomial.logcdf
-        if np.ndim(value):
-            raise TypeError(
-                f"ZeroInflatedNegativeBinomial.logcdf expects a scalar value but received a {np.ndim(value)}-dimensional object."
-            )
-
         return bound(
             logaddexp(at.log1p(-psi), at.log(psi) + _logcdf(nbinom, value, {}, n, p)),
             0 <= value,